First break down of code into modular files using webpack for compiling
