+++ css/lists.css
@@ -1,0 +1,42 @@
+.list-group.flex-wrap {
+    border-left: 1px solid rgba(0,0,0,.125);
+    border-top: 1px solid rgba(0,0,0,.125);
+}
+.list-group.flex-wrap .list-group-item {
+    border-radius: 0;
+    border: 1px solid rgba(0,0,0,.125);
+    border-left: none;
+    border-top: none;
+}
+
+.input-group.list-group-item {
+    display: flex;
+    padding: 0;
+}
+
+.input-group.list-group-item .form-control {
+    background: transparent;
+}
+
+.input-group.list-group-item .form-control,
+.input-group.list-group-item .input-group-text,
+.input-group.list-group-item .btn {
+    border: none;
+    border-radius: 0;
+    background: transparent;
+}
+.input-group.list-group-item .btn-outline-secondary:hover,
+.input-group.list-group-item .btn-outline-secondary:focus {
+    color: #212529;
+}
+
+.list-group.list-disabled {
+    border-left: 1px solid rgba(0,0,0,.0625);
+    border-top: 1px solid rgba(0,0,0,.0625);
+}
+.list-disabled .list-group-item {
+    opacity: .5;
+}
+.list-disabled .list-group-item:hover {
+    opacity: 1;
+}
