+++ package.json
@@ -1,0 +1,34 @@
+{
+  "name": "instructor-utilities",
+  "version": "1.0.0",
+  "description": "Tools and utilities built for helping assist instructing bootcamp classes",
+  "private": true,
+  "scripts": {
+    "watch": "webpack --hide-modules --mode=development --watch",
+    "production": "webpack --hide-modules --mode=production"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ac524/instructor-utilities.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/ac524/instructor-utilities/issues"
+  },
+  "homepage": "https://github.com/ac524/instructor-utilities#readme",
+  "devDependencies": {
+    "@babel/core": "^7.9.6",
+    "@babel/polyfill": "^7.8.7",
+    "@babel/preset-env": "^7.9.6",
+    "autoprefixer": "^9.8.0",
+    "babel-loader": "^8.1.0",
+    "css-loader": "^3.5.3",
+    "mini-css-extract-plugin": "^0.9.0",
+    "postcss-loader": "^3.0.0",
+    "sass-loader": "^8.0.2",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.11"
+  }
+}
